refactor: SOF-1090 move `findGraphicsSetup` to tv2-common
