+++ src/tv2_afvd_showstyle/helpers/config.ts
@@ -4,7 +4,7 @@ import {
 	IShowStyleContext,
 	TableConfigItemValue
 } from '@tv2media/blueprints-integration'
-import { TableConfigGraphicsSetup, TV2ShowstyleBlueprintConfigBase } from 'tv2-common'
+import { findGraphicsSetup, TableConfigGraphicsSetup, TV2ShowstyleBlueprintConfigBase } from 'tv2-common'
 import { BlueprintConfig as BlueprintConfigBase } from '../../tv2_afvd_studio/helpers/config'
 
 export interface GalleryTableConfigGraphicsSetup extends TableConfigGraphicsSetup {
@@ -22,25 +22,14 @@ export interface ShowStyleConfig extends TV2ShowstyleBlueprintConfigBase {
 	GraphicsSetups: GalleryTableConfigGraphicsSetup[]
 }
 
-function findGraphicsSetup(context: ICommonContext, config: ShowStyleConfig): GalleryTableConfigGraphicsSetup {
-	const foundTableConfigGraphicsSetup: GalleryTableConfigGraphicsSetup | undefined = config.GraphicsSetups.find(
-		tableConfigGraphicsSetup => tableConfigGraphicsSetup.Name === config.SelectedGraphicsSetupName
-	)
-	if (!foundTableConfigGraphicsSetup) {
-		context.logWarning(`No graphics setup found for profile: ${config.SelectedGraphicsSetupName}`)
-		return {
-			Name: '',
-			VcpConcept: '',
-			OvlShowName: '',
-			FullShowName: ''
-		}
-	}
-	return foundTableConfigGraphicsSetup
-}
-
 export function parseConfig(context: ICommonContext, rawConfig: IBlueprintConfig): any {
 	const showstyleConfig = (rawConfig as unknown) as ShowStyleConfig
-	const selectedGraphicsSetup = findGraphicsSetup(context, showstyleConfig)
+	const selectedGraphicsSetup = findGraphicsSetup(context, showstyleConfig, {
+		Name: '',
+		VcpConcept: '',
+		OvlShowName: '',
+		FullShowName: ''
+	})
 	return {
 		showStyle: showstyleConfig,
 		selectedGraphicsSetup
